requirements/requirements.txt
# Azure Storage and Identity dependencies
azure-storage-blob==12.19.0
azure-identity==1.15.0
azure-eventgrid==4.18.0

# Core dependencies
requests>=2.28.0
typing-extensions>=4.0.0
python-dotenv>=1.0.0

# PDF processing dependencies
opencv-python-headless==4.8.1.78
deskew>=1.3.3
numpy>=1.24.0,<2.0.0
scipy>=1.11.0
scikit-image>=0.21.0
Pillow>=10.0.0

# Optional: Additional PDF processing libraries
PyPDF2==3.0.1
pdfplumber==0.10.0

# OCR dependencies (choose one or more)
# Optional OCR engines
pytesseract>=0.3.10
easyocr>=1.7.0  # Optional, advanced OCR

# Document Classification Dependencies
scikit-learn>=1.3.0
# Optional ML dependencies for advanced classification
sentence-transformers>=2.2.0  # For advanced text embeddings (optional)
transformers>=4.21.0          # For transformer-based classification (optional)

# Azure Document Intelligence OCR
azure-ai-documentintelligence>=1.0.0b1  # Azure Document Intelligence v4
azure-core>=1.29.0            # Azure core libraries
azure-identity>=1.15.0        # Azure authentication (already included)

# Additional OCR and document processing
Pillow>=10.0.0               # Image processing (PIL)
io                           # Built-in I/O operations
base64                       # Built-in base64 encoding

# Field extraction and validation
python-dateutil>=2.8.2       # Date parsing and normalization
phonenumbers>=8.13.0         # Phone number validation and formatting
email-validator>=2.1.0       # Email address validation
validators>=0.22.0           # General data validation

# Optional Service Bus for HITL queue
azure-servicebus>=7.11.0     # Azure Service Bus for enterprise queuing (optional)

# HITL Review Application
fastapi>=0.104.0             # Web framework for HITL review interface
uvicorn>=0.24.0              # ASGI server for FastAPI
httpx>=0.25.0                # HTTP client for Teams notifications
jinja2>=3.1.0                # Template engine for web interface
joblib>=1.3.0
pandas>=1.5.0  # For data handling
matplotlib>=3.6.0  # For visualization (optional)
# paddleocr>=2.7.0     # PaddleOCR (uncomment if needed)

# Machine Learning (optional for advanced features)
# scikit-learn>=1.3.0
# tensorflow>=2.13.0

# Note: Tesseract binary installation required for pytesseract
# Ubuntu/Debian: sudo apt-get install tesseract-ocr
# macOS: brew install tesseract
# Windows: Download from https://github.com/UB-Mannheim/tesseract/wiki